impl simple table scan

+++ src/allocator.rs
@@ -1,16 +1,16 @@
 pub mod allocator {
     #[derive(Clone)]
     pub struct Allocator {
-        pub table_id: u64,
-        pub base: u64,
+        pub table_id: usize,
+        pub base: usize,
     }
 
     impl Allocator {
-        pub fn new(table_id: u64) -> Box<Allocator> {
+        pub fn new(table_id: usize) -> Box<Allocator> {
             Box::new(
                 Allocator {
                     table_id: table_id,
-                    base: 1 as u64,
+                    base: 1 as usize,
                 }
             )
         }
